Split Classes. Calculate age from date of birth and date of death. Added methods for parents and toString
