+++ src/Person.java
@@ -1,0 +1,47 @@
+import java.time.LocalDate;
+import java.time.Period;
+import java.util.ArrayList;
+import java.util.List;
+
+public class Person {
+    private String name;
+    private LocalDate birthDate;
+    private LocalDate deathDate;
+    private List<Person> parents;
+    private List<Person> children;
+
+    public Person(String name, LocalDate birthDate, LocalDate deathDate) {
+        this.name = name;
+        this.birthDate = birthDate;
+        this.deathDate = deathDate;
+        this.parents = new ArrayList<>();
+        this.children = new ArrayList<>();
+    }
+
+    public void addChild(Person child) {
+        this.children.add(child);
+        child.parents.add(this);
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public List<Person> getChildren() {
+        return children;
+    }
+
+    public List<Person> getParents() {
+        return parents;
+    }
+
+    public int getAge() {
+        LocalDate endDate = (deathDate != null) ? deathDate : LocalDate.now();
+        return Period.between(birthDate, endDate).getYears();
+    }
+
+    @Override
+    public String toString() {
+        return name + ", " + getAge() + " years old";
+    }
+}
